requirements.txt
flask
flask-sqlalchemy
flask-smorest
dataclasses
python-dotenv
flask-jwt-extended
werkzeug